Remove test double duplication
Extract TestReaderWriter double to its own file and import in HTTPServerTest
and EchoServerTest. Rename ReadableWriteable test double in MyServerTest
to reflect its logging responsibility.
