+++ src/test/java/server/EchoServerTest.java
@@ -3,7 +3,6 @@ package server;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -31,38 +30,5 @@ public class EchoServerTest {
                         "first line\n",
                         "second line\n"),
                 testReaderWriter.received());
-    }
-
-    private static class TestReaderWriter implements ReadableWriteable {
-
-        private final List<String> toRead = new ArrayList<>();
-        private final List<String> written = new ArrayList<>();
-
-        @Override
-        public String readLine() {
-            if (toRead.isEmpty()) {
-                return null;
-            }
-            return toRead.remove(0);
-        }
-
-        @Override
-        public void writeLine(String message) {
-            written.add(message);
-        }
-
-        @Override
-        public void close() {
-
-        }
-
-        public TestReaderWriter send(String message) {
-            toRead.add(message);
-            return this;
-        }
-
-        public List<String> received() {
-            return written;
-        }
     }
 }
